requirement.txt
google-genai
ultralytics
transformers
torch
pillow
fpdf
opencv-python
numpy